requirements.txt
PyQt5
pandas
xgboost
ezdxf
joblib
python-dotenv